remove 2D eigenvalues compile unit
